refactor(admin): 관리자 화면 borderless 구조 정리
